Adding more data to films

+++ schema/cinema-schema.sql
@@ -5,7 +5,7 @@ USE cinema;
 CREATE TABLE films(
     film_id INT AUTO_INCREMENT,
     film_type VARCHAR(40) NOT NULL,
-    price INT,
+    price DECIMAL(10,2),
     rating INT,
     screen_no INT,
     PRIMARY KEY(film_id)
